requirements.txt
torch>=2.0.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
scipy>=1.7.0
tqdm>=4.60.0
gymnasium>=0.26.0
stable-baselines3>=1.6.0
pyyaml>=6.0.0